move search endpoint, handler, view into its own folder, #326
