+++ server/routes/generic.js
@@ -2,7 +2,6 @@
 
 var serveFile = require('../handlers/serve-file.js');
 var landing = require('../handlers/landing/landing.js');
-var search = require('../handlers/search.js')
 
 module.exports = [
   {
@@ -15,10 +14,5 @@ module.exports = [
     path: '/{path*}',
     handler: serveFile,
     config: { auth: false }
-  },
-  {
-    method: 'GET',
-    path: '/search',
-    handler: search
   }
 ];
