+++ server/routes/search.js
@@ -1,0 +1,11 @@
+'use strict';
+
+var search = require('../handlers/search/search.js');
+
+module.exports = [
+  {
+    method: 'GET',
+    path: '/search',
+    handler: search
+  }
+];
